Separated MdisNacSensorModelTest into cpp and header. Modified runTests to take a second argument (for a test data directory.) Defined a root cmake variable to point to test data area.
